Move Edit on GitHub links to non-cached footer
A small side-effect, which I think we should accept, is that the custom, dynamic content is rendered before the fixed one.

The minimal-mistakes theme offers a dynamic as well as a fully-cached footer include.

The cached one would have the same page.path, page.url parameters. The "footer/custom.html" allows customization, with dynamic data.

More details and background:

The inclusion in the theme:
https://github.com/mmistakes/minimal-mistakes/blob/93022d6432f277eb5d83677fa50e4120f1281b04/_layouts/default.html#L29